requirements.txt
flask
openai
python-dotenv
sqlalchemy
psycopg2-binary
pymysql
cryptography